Add dossier to fragment, update & refactor
