+++ ebl/dossiers/application/dossiers_repository.py
@@ -8,7 +8,7 @@ from ebl.dossiers.domain.dossier_record import (
 
 class DossiersRepository(ABC):
     @abstractmethod
-    def query_by_ids(self, ids: Sequence[str]) -> DossierRecord: ...
+    def query_by_ids(self, ids: Sequence[str]) -> Sequence[DossierRecord]: ...
 
     @abstractmethod
     def create(self, dossier_record: DossierRecord) -> str: ...
